+++ app/src/main/java/org/stepik/android/exams/core/interactor/contacts/NavigationInteractor.kt
@@ -4,6 +4,6 @@ import io.reactivex.Observable
 import org.stepik.android.exams.data.model.LessonTheoryWrapper
 
 interface NavigationInteractor {
-    fun resolveNextLesson(topicId: String, lesson: Int, move: Boolean): Observable<List<LessonTheoryWrapper>>
-    fun resolvePrevLesson(topicId: String, lesson: Int, move: Boolean): Observable<List<LessonTheoryWrapper>>
+    fun resolveNextLesson(topicId: String, lesson: Long, move: Boolean): Observable<List<LessonTheoryWrapper>>
+    fun resolvePrevLesson(topicId: String, lesson: Long, move: Boolean): Observable<List<LessonTheoryWrapper>>
 }
